+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Re:Emerald Changelog
 
 ## In-development
+**Update 22**
+- Removed the VS Seeker behavior, as it was causing issues with the Match Call rematch system
+- Improved rematch system slightly (Now the steps left to check for a rematch will only reset if a new rematch was triggered. Previously, it would also reset if there was an active rematch on the route you entered)
+
 **Update 21**
 - Added the sidequests to obtain the second bike from Rydel.